fastapi-server/pyproject.toml
[tool.poetry]
name = "main"
version = "1.0.0"
description = "FastAPI starter template with caching and rate-limiting built-in"
authors = []
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
anyio = "^3.6.2"
click = "^8.1.3"
h11 = "^0.14.0"
httptools = "^0.5.0"
idna = "^3.4"
pydantic = "^1.10.4"
python-dotenv = "^1.0.0"
PyYAML = "^6.0"
sniffio = "^1.3.0"
uvicorn = "^0.20.0"
uvloop = "^0.17.0"
watchfiles = "^0.18.1"
websockets = "^10.4"
openai = "^1.1.0"
sseclient = "^0.0.27"
sseclient-py = "^1.7.2"
langchain = "0.0.303"
ipykernel = "^6.23.1"
supabase = "^1.0.3"
google-search-results = "^2.4.2"
pandas = "^2.1.0"
unstructured = "^0.10.12"
twilio = "^8.8.0"
metaphor-python = "^0.1.16"
polygon-api-client = "^1.12.4"
yfinance = "^0.2.31"
psycopg2-binary = "^2.9.9"
pytrends = "^4.9.2"
newsapi-python = "^0.2.7"
portkey-ai = "^1.1.5"
llama-index = "^0.10.18"
llama-index-agent-openai = "^0.1.5"
fastapi = "^0.110.1"
tiktoken = "^0.6.0"
llama-index-tools-tavily-research = "^0.1.3"
llama-index-tools-database = "^0.1.3"
llama-index-llms-openai = "^0.1.14"
modal = "^0.62.43"
redis = "^5.0.3"
fastapi-cache2 = "^0.2.1"
fastapi-limiter = "^0.1.6"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
dev = "main:dev"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
